requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
astunparse==1.6.3
boto3==1.35.20
botocore==1.35.20
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.114.2
flatbuffers==24.3.25
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.2
h11==0.14.0
h5py==3.12.1
idna==3.10
jmespath==1.0.1
keras==3.6.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nibabel==5.2.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.5
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.32.3
rich==13.9.2
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-addons==0.23.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
Werkzeug==3.0.4
wrapt==1.16.0
